pyproject.toml
[tool.poetry]
name = "fantastic-pub-sub"
version = "0.1.0"
description = ""
authors = ["Claudio Cilione <uglywizard@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-pubsub = "2.2.0"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
